moved remove exercise function to exercise card
